+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/Views/ViewBase.cs
@@ -34,7 +34,8 @@ namespace UnityEngine.InputSystem.Editor
         {
             if (m_ViewStateSelector == null)
             {
-                Debug.LogWarning($"View '{GetType().Name}' has no selector and will not render. Create a selector for the " +
+                Debug.LogWarning(
+                    $"View '{GetType().Name}' has no selector and will not render. Create a selector for the " +
                     $"view using the CreateSelector method.");
                 return;
             }
@@ -113,11 +114,15 @@ namespace UnityEngine.InputSystem.Editor
         private IList<IView> m_ChildViews;
         private bool m_IsFirstUpdate = true;
 
+
+        protected bool IsDuplicateShortcutPressed(KeyDownEvent keyDownEvent)
+        {
 #if UNITY_STANDALONE_OSX
-        protected bool IsCommandOrControlPressed(KeyDownEvent keyDownEvent) => keyDownEvent.keyCode == KeyCode.D && keyDownEvent.modifiers == EventModifiers.Command;
+            return keyDownEvent.keyCode == KeyCode.D && keyDownEvent.modifiers == EventModifiers.Command;
 #else
-        private bool IsCommandOrControlPressed(KeyDownEvent keyDownEvent) => keyDownEvent.keyCode == KeyCode.D && keyDownEvent.modifiers == EventModifiers.Control;
+            return keyDownEvent.keyCode == KeyCode.D && keyDownEvent.modifiers == EventModifiers.Control;
 #endif
+        }
     }
 
     internal class ViewStateSelector<TReturn> : IViewStateSelector<TReturn>